Backend separate file restructure
