+++ server/package.json
@@ -28,7 +28,10 @@
   "devDependencies": {
     "@types/express": "^4.17.17",
     "@types/jest": "^29.5.0",
+    "@types/lodash": "^4.14.191",
     "@types/node": "^18.15.5",
+    "@types/nodemailer": "^6.4.7",
+    "@types/uuid": "^9.0.1",
     "@typescript-eslint/parser": "^5.56.0",
     "dotenv": "^16.0.3",
     "ts-node": "^10.9.1",
