--- server/src/userfunctions.ts
@@ -1,1315 +1,0 @@
-const couchdbUrl = (process.env.COUCHDB_URL == undefined) ? "" : process.env.COUCHDB_URL.endsWith("/") ? process.env.COUCHDB_URL.slice(0,-1): process.env.COUCHDB_URL;
-const couchDatabase = (process.env.COUCHDB_DATABASE == undefined) ? "" : process.env.COUCHDB_DATABASE;
-const couchKey = process.env.COUCHDB_HMAC_KEY;
-const couchAdminUser = process.env.COUCHDB_ADMIN_USER;
-const couchAdminPassword = process.env.COUCHDB_ADMIN_PASSWORD;
-const groceryUrl = (process.env.GROCERY_URL == undefined) ? "" : process.env.GROCERY_URL.endsWith("/") ? process.env.GROCERY_URL.slice(0,-1): process.env.GROCERY_URL;
-const groceryAPIUrl = (process.env.GROCERY_API_URL == undefined) ? "" : process.env.GROCERY_API_URL.endsWith("/") ? process.env.GROCERY_API_URL.slice(0,-1): process.env.GROCERY_API_URL;
-const smtpHost = process.env.SMTP_HOST;
-const smtpPort = process.env.SMTP_PORT;
-const smtpSecure = Boolean(process.env.SMTP_SECURE);
-const smtpUser = process.env.SMTP_USER;
-const smtpPassword = process.env.SMTP_PASSWORD;
-const smtpFrom = process.env.SMTP_FROM;
-const enableScheduling = (process.env.ENABLE_SCHEDULING == undefined) ? true : Boolean(process.env.ENABLE_SCHEDULING);
-const resolveConflictsFrequencyMinutes = (process.env.RESOLVE_CONFLICTS_FREQUENCY_MINUTES == undefined) ? 15 : process.env.RESOLVE_CONFLICTS_FREQUENCY_MINUTES;
-const expireJWTFrequencyMinutes = (process.env.EXPIRE_JWT_FREQUENCY_MINUTES == undefined) ? 10 : process.env.EXPIRE_JWT_FREQUENCY_MINUTES;
-const couchStandardRole = "crud";
-const couchAdminRole = "dbadmin";
-const couchUserPrefix = "org.couchdb.user";
-const conflictsViewID = "_conflicts_only_view_id";
-const conflictsViewName = "conflicts_view";
-const smtpOptions = {
-    host: smtpHost, port: smtpPort, 
-    auth: { user: smtpUser, pass: smtpPassword}
-};
-const nodemailer = require('nodemailer');
-const jose = require('jose');
-const axios = require('axios');
-const nanoAdmin = require('nano');
-const nanoAdminOpts = {
-    url: couchdbUrl,
-    requestDefaults: {
-        headers: { Authorization: "Basic "+ Buffer.from(couchAdminUser+":"+couchAdminPassword).toString('base64') }
-    }
-}
-let todosNanoAsAdmin = nanoAdmin(nanoAdminOpts);
-let usersNanoAsAdmin = nanoAdmin(nanoAdminOpts);
-let todosDBAsAdmin;
-let usersDBAsAdmin;
-const _ = require('lodash');
-const { v4: uuidv4 } = require('uuid');
-const { cloneDeep, isEmpty, isEqual, isSafeInteger } = require('lodash');
-const {  emailPatternValidation, usernamePatternValidation, fullnamePatternValidation,
-     uomContent, globalItems, categories } = require('./utilities');
-let uomContentVersion = 0;
-const targetUomContentVersion = 3;
-let categoriesVersion = 0;
-const targetCategoriesVersion = 1;
-let globalItemVersion = 0;
-const targetGlobalItemVersion = 1;
-let schemaVersion = 0;
-const targetSchemaVersion = 3;
-const refreshTokenExpires = (process.env.REFRESH_TOKEN_EXPIRES == undefined) ? "30d" : process.env.REFRESH_TOKEN_EXPIRES;
-const accessTokenExpires = (process.env.ACCESS_TOKEN_EXPIRES == undefined) ? "5m" : process.env.ACCESS_TOKEN_EXPIRES;
-const JWTKey = new TextEncoder().encode(couchKey);
-
-async function couchLogin(username, password) {
-    const loginResponse = {
-        loginSuccessful: true,
-        loginRoles: []
-    }
-    const config = {
-        method: 'get',
-        url: couchdbUrl+"/_session",
-        auth: { username: username, password: password},
-        responseType: 'json'
-    }
-    let res=null;
-    try  {res = await axios(config)}
-    catch(err) {loginResponse.loginSuccessful = false};
-    if (loginResponse.loginSuccessful) {
-        if (res.status != 200) {
-            loginResponse.loginSuccessful = false;
-        }
-        if (loginResponse.loginSuccessful && (res.data?.ok != true)) {
-            loginResponse.loginSuccessful = false;
-        }
-    }
-    if (loginResponse.loginSuccessful) {
-        loginResponse.loginRoles = res.data.userCtx.roles;
-    }
-    return(loginResponse);
-
-}
-
-function getNested(obj, ...args) {
-    return args.reduce((obj, level) => obj && obj[level], obj)
-  }
-
-async function doesDBExist() {
-    let retrieveError = false;
-    let res = null;
-    try { res = await todosNanoAsAdmin.db.get(couchDatabase)}
-    catch(err) { retrieveError = true }
-    if (retrieveError || res == null) {
-        console.log("ERROR: could not retrieve database info.");
-        return (false);
-    } else {
-        return (true)
-    }
-}
-
-async function createDB() {
-    let createError = false;
-    try { await todosNanoAsAdmin.db.create(couchDatabase)}
-    catch(err) {  createError = true }
-    if (createError) return (false);
-    console.log("STATUS: Initiatialization, Database "+couchDatabase+" created.");
-    return (createError);
-}
-
-async function createDBIfNotExists() {
-    let dbCreated=false
-    if (!(await doesDBExist())) {
-        dbCreated=await createDB()
-    }
-    return (dbCreated)
-}
-
-async function totalDocCount(db) {
-    const info = await db.info();
-    return info.doc_count;
-}
-
-async function setDBSecurity() {
-    let errorSettingSecurity = false;
-    let config = {
-        method: 'get',
-        url: couchdbUrl+"/"+couchDatabase+"/_security",
-        auth: {username: couchAdminUser, password: couchAdminPassword},
-        responseType: 'json'
-    }
-    let res = null;
-    try { res = await axios(config)}
-    catch(err) { console.log("ERROR setting security:",err); errorSettingSecurity= true }
-    if (errorSettingSecurity) return (!errorSettingSecurity);
-    let newSecurity = _.cloneDeep(res.data);
-    let securityNeedsUpdated = false;
-    if ((getNested(res.data.members.roles.length) == 0) || (getNested(res.data.members.roles.length) == undefined)) {
-        newSecurity.members.roles = [couchStandardRole];
-        securityNeedsUpdated = true;
-    } else {
-        if (!res.data.members.roles.includes(couchStandardRole)) {
-            newSecurity.members.roles.push(couchStandardRole);
-            securityNeedsUpdated = true;
-        }
-    }
-    if ((getNested(res.data.admins.roles.length) == 0) || (getNested(res.data.admins.roles.length) == undefined)) {
-        newSecurity.admins.roles = [couchAdminRole];
-        securityNeedsUpdated = true;
-    } else {
-        if (!res.data.admins.roles.includes(couchAdminRole)) {
-            newSecurity.admins.roles.push(couchAdminRole);
-            securityNeedsUpdated = true;
-        }
-    }
-    if (!securityNeedsUpdated) {
-        console.log("STATUS: Security roles set correctly");
-        return (true);
-    }
-    let configSec = {
-        method: 'put',
-        url: couchdbUrl+"/"+couchDatabase+"/_security",
-        auth: {username: couchAdminUser, password: couchAdminPassword},
-        responseType: 'json',
-        data: newSecurity
-    }
-    errorSettingSecurity = false;
-    try { res = await axios(configSec)}
-    catch(err) { console.log("ERROR setting security:", err); errorSettingSecurity = true }
-    if (errorSettingSecurity) {
-        console.log("ERROR: Problem setting database security")
-    } else {
-        console.log("STATUS: Database security roles added");
-    }
-    return (!errorSettingSecurity);
-}
-
-async function getLatestDBUUIDDoc() {
-    const dbidq = {
-        selector: { type: { "$eq": "dbuuid" }}
-    }
-    let foundIDDocs = null;
-    try {foundIDDocs =  await todosDBAsAdmin.find(dbidq);}
-    catch(err) {console.log("ERROR: could not read dbUUID record"); return undefined;}
-    let foundIDDoc = undefined;
-    if (foundIDDocs && foundIDDocs.hasOwnProperty('docs')) {
-        if (foundIDDocs.docs.length > 0) {foundIDDoc = foundIDDocs.docs[0]}
-    }   
-    return foundIDDoc;
-}
-
-async function updateDBUUIDDoc(dbuuidDoc) {
-    dbuuidDoc.updatedAt = (new Date()).toISOString();
-    let dbResp = null;
-    try {dbResp = await todosDBAsAdmin.insert(dbuuidDoc);}
-    catch(err) {console.log("ERROR: could not update dbUUID record:",JSON.stringify(err)); return null;}
-    return dbResp;
-}
-
-async function addDBIdentifier() {
-    let foundIDDoc = await getLatestDBUUIDDoc();
-    if (foundIDDoc == undefined) {
-        const newDoc = {
-            type: "dbuuid",
-            name: "Database UUID",
-            "uuid": uuidv4(),
-            "uomContentVersion": 0,
-            "categoriesVersion": 0,
-            "globalItemVersion": 0,
-            "schemaVersion": 0,
-            updatedAt: (new Date()).toISOString()
-        }
-        let dbResp=await updateDBUUIDDoc(newDoc);
-        if (dbResp != null) {console.log("STATUS: UUID created in DB: ", newDoc.uuid)}  
-    } else {
-        if (!foundIDDoc.hasOwnProperty("uuid")) {
-            console.log("ERROR: Database UUID doc exists without uuid. Please correct and restart.");
-            return false;
-        }
-        if (!foundIDDoc.hasOwnProperty("uomContentVersion")) {
-            foundIDDoc.uomContentVersion = 0;
-            let dbResp = await updateDBUUIDDoc(foundIDDoc);
-            if (dbResp == null) { console.log("ERROR: updating UUID record with uomContentVersion");  } 
-            else { console.log("STATUS: Updated UOM Content Version, was missing.") }
-        } else {
-            uomContentVersion = foundIDDoc.uomContentVersion;
-        }
-        foundIDDoc = await getLatestDBUUIDDoc();
-        if (!foundIDDoc.hasOwnProperty("categoriesVersion")) {
-            foundIDDoc.categoriesVersion = 0;
-            let dbResp = await updateDBUUIDDoc(foundIDDoc);
-            if (dbResp == null) { console.log("ERROR: updating UUID record with categoriesVersion: ",dbResp);  } 
-            else { console.log("STATUS: Updated Categories Content Version, was missing.") }
-        } else {
-            categoriesVersion = foundIDDoc.categoriesVersion;
-        }
-        foundIDDoc = await getLatestDBUUIDDoc();        
-        if (!foundIDDoc.hasOwnProperty("globalItemVersion")) {
-            foundIDDoc.globalItemVersion = 0;
-            let dbResp = await updateDBUUIDDoc(foundIDDoc);
-            if (dbResp == null) { console.log("ERROR: updating UUID record with globalItemVersion");  } 
-            else { console.log("STATUS: Updated global Item Content Version, was missing."); }
-        } else {
-            globalItemVersion = foundIDDoc.globalItemVersion;
-        }
-        foundIDDoc = await getLatestDBUUIDDoc();
-        if (!foundIDDoc.hasOwnProperty("schemaVersion")) {
-            foundIDDoc.schemaVersion = 0;
-            let dbResp = await updateDBUUIDDoc(foundIDDoc);
-            if (dbResp == null) { console.log("ERROR: updating UUID record with schemaVersion") } 
-            else { console.log("STATUS: Updated Categories Content Version, was missing.");  }
-        } else {
-            schemaVersion = foundIDDoc.schemaVersion;
-        }
-    }
-}
-
-async function createUOMContent() {
-    const dbuomq = {
-        selector: { type: { "$eq": "uom" }},
-        limit: await totalDocCount(todosDBAsAdmin)
-    }
-    let foundUOMDocs =  await todosDBAsAdmin.find(dbuomq);
-    for (let i = 0; i < uomContent.length; i++) {
-        let uom = uomContent[i];
-        const docIdx=foundUOMDocs.docs.findIndex((el) => (el.name.toUpperCase() === uom.name.toUpperCase() || el._id === uom._id));
-        let needsAdded=true;
-        if (docIdx !== -1) {
-            let thisDoc = foundUOMDocs.docs[docIdx];
-            if (thisDoc._id === uom._id) {
-                console.log("STATUS: UOM ",uom.name," already exists...skipping...");
-                needsAdded=false;
-            } else {
-                let dbResp = null;
-                try { dbResp = await todosDBAsAdmin.destroy(thisDoc._id,thisDoc._rev)}
-                catch(err) {console.log("ERROR deleting / replacing existing UOM: ", err);}
-            }
-        }
-        if (needsAdded) {
-            console.log("STATUS: Adding uom ",uom.name, " ", uom.description);
-            let dbResp = null;
-            try { dbResp = await todosDBAsAdmin.insert(uom);}
-            catch(err) { console.log("ERROR: adding uom ",uom.uom, " error: ",err);}
-        } else {
-            console.log("STATUS: UOM ",uom.name," already exists...skipping");
-        }
-    };
-    console.log("STATUS: Finished adding units of measure, updating to UOM Content Version:",targetUomContentVersion);
-    let foundIDDoc = await getLatestDBUUIDDoc();
-    if (foundIDDoc == undefined) {
-        console.log("ERROR: Couldn't update database content version record.");
-    } else {
-        foundIDDoc.uomContentVersion = targetUomContentVersion;
-        let dbResp = await updateDBUUIDDoc(foundIDDoc);
-        if (dbResp == null) { console.log("ERROR Couldn't update UOM target version.")}
-        else { console.log("STATUS: Updated UOM Target Version successfully."); }
-    }
-}
-
-async function createCategoriesContent() {
-    const dbcatq = {
-        selector: { type: { "$eq": "category" }},
-        limit: await totalDocCount(todosDBAsAdmin)
-    }
-    let foundCategoryDocs =  await todosDBAsAdmin.find(dbcatq);
-    for (let i = 0; i < categories.length; i++) {
-        let category = categories[i];
-        category.type = "category";
-        category.color = "#ffffff";
-        const docIdx=foundCategoryDocs.docs.findIndex((el) => (el.name.toUpperCase() === category.name.toUpperCase() || el._id === category._id));
-        let needsAdded = true;
-        if (docIdx !== -1) {
-            let thisDoc = foundCategoryDocs.docs[docIdx]
-            if (thisDoc._id === category._id) {
-                console.log("STATUS: Category ",category.name," already exists...skipping");
-                needsAdded=false;
-            } else {
-                let dbResp = null;
-                try { dbResp = await todosDBAsAdmin.destroy(thisDoc._id,thisDoc._rev)}
-                catch(err) { console.log("ERROR: deleting category for replacement:", err);}
-            }
-        }
-        if (needsAdded) {
-            console.log("STATUS: Adding category ",category.name);
-            let dbResp = null;
-            try { dbResp = await todosDBAsAdmin.insert(category);}
-            catch(err) { console.log("ERROR: adding category ",category.name, " error: ",err);}
-        } 
-    };
-    console.log("STATUS: Finished adding categories, updating to category Version:",targetCategoriesVersion);
-    let foundIDDoc = await getLatestDBUUIDDoc();
-    if (foundIDDoc == undefined) {
-        console.log("ERROR: Couldn't update database content version record.");
-    } else {
-        foundIDDoc.categoriesVersion = targetCategoriesVersion;
-        let dbResp = null;
-        try { dbResp = await todosDBAsAdmin.insert(foundIDDoc)}
-        catch(err) { console.log("ERROR: Couldn't update Categories target version.")};
-        console.log("STATUS: Updated Categories Target Version successfully.");
-    }
-}
-
-async function createGlobalItemContent() {
-    const dbglobalq = {
-        selector: { type: { "$eq": "globalitem" }},
-        limit: await totalDocCount(todosDBAsAdmin)
-    }
-    let foundGlobalItemDocs =  await todosDBAsAdmin.find(dbglobalq);
-    for (let i = 0; i < globalItems.length; i++) {
-        let globalItem = globalItems[i];
-        globalItem.type = "globalitem";
-        const docIdx=foundGlobalItemDocs.docs.findIndex((el) => el.name === globalItem.name );
-        if (docIdx == -1) {
-            console.log("STATUS: Adding global item ",globalItem.name);
-            let dbResp = null;
-            try { dbResp = await todosDBAsAdmin.insert(globalItem);}
-            catch(err) { console.log("ERROR: adding global item ",globalItem.name, " error: ",err);}
-        } else {
-            console.log("STATUS: Global Item ",globalItem.name," already exists...skipping");
-        }
-    };
-    console.log("STATUS: Finished adding global Items, updating to Global Item Version:",targetGlobalItemVersion);
-    let foundIDDoc = await getLatestDBUUIDDoc();
-    if (foundIDDoc == undefined) {
-        console.log("ERROR: Couldn't update database content version record.");
-    } else {
-        foundIDDoc.globalItemVersion = targetGlobalItemVersion;
-        let dbResp = await updateDBUUIDDoc(foundIDDoc);
-        if (dbResp == null) { console.log("ERROR: Couldn't update Global Item target version.") }
-        else {console.log("STATUS: Updated Global Item Target Version successfully.");}
-    }
-}
-
-async function checkAndCreateContent() {
-    console.log("STATUS: Current UOM Content Version:",uomContentVersion," Target Version:",targetUomContentVersion);
-    if (uomContentVersion === targetUomContentVersion) {
-        console.log("STATUS: At current version, skipping UOM Content creation");
-    } else {
-        console.log("STATUS: Creating UOM Content...");
-        await createUOMContent();
-    }
-    console.log("STATUS: Current Category Content Version:",categoriesVersion," Target Version:", targetCategoriesVersion);
-    if (categoriesVersion === targetCategoriesVersion) {
-        console.log("STATUS: At current category version, skipping category creation");
-    } else {
-        console.log("STATUS: Creating category Content...");
-        await createCategoriesContent();
-    }
-    console.log("STATUS: Current Global Item Content Version:",globalItemVersion," Target Version:", targetGlobalItemVersion);
-    if (globalItemVersion === targetGlobalItemVersion) {
-        console.log("STATUS: At current Global item version, skipping global item creation");
-    } else {
-        console.log("STATUS: Creating Global Item Content...");
-        await createGlobalItemContent();
-    }
-}
-
-async function addStockedAtIndicatorToSchema() {
-    let updateSuccess = true;
-    console.log("STATUS: Upgrading schema to support stocked at indicators.");
-    const itemq = { selector: { type: { "$eq": "item"}},
-                    limit: await totalDocCount(todosDBAsAdmin)};
-    let foundItemDocs;
-    try {foundItemDocs = await todosDBAsAdmin.find(itemq);}                
-    catch(err) {console.log("ERROR: Could not find item docs to update during schema update"); return false;}
-    console.log("STATUS: Found items to update :", foundItemDocs.docs.length)
-    for (let i = 0; i < foundItemDocs.docs.length; i++) {
-        const foundItemDoc = foundItemDocs.docs[i];
-        console.log("Processing item: ", foundItemDoc.name);
-        let docChanged = false;
-        if (foundItemDoc.hasOwnProperty("lists")) {
-            for (let j = 0; j < foundItemDoc.lists.length; j++) {
-                console.log("list: ",JSON.stringify(foundItemDoc.lists[j]));
-                if (!foundItemDoc.lists[j].hasOwnProperty("stockedAt")) {
-                    console.log("Didn't have stockedAt property, adding...");
-                    foundItemDoc.lists[j].stockedAt = true;
-                    docChanged = true;
-                }
-            }
-        }
-        if (docChanged) {
-            let dbResp = null;
-            try { dbResp = await todosDBAsAdmin.insert(foundItemDoc)}
-            catch(err) { console.log("ERROR: Couldn't update item with stocked indicator.");
-                         updateSuccess = false;}
-        }
-    }
-    return updateSuccess;
-}
-
-async function restructureListGroupSchema() {
-    let updateSuccess = true;
-    console.log("STATUS: Upgrading schema to support listGroups. Most data will be lost in this upgrade.");
-    // Delete both lists and items because of structure changes and because categories in list are most likely
-    // completely replaced with new Category content/system IDs at same time
-    const delq = { selector: { type : { "$in": ["list","item"]}}, limit: await totalDocCount(todosDBAsAdmin)};
-    let foundDelDocs;
-    try { foundDelDocs = await todosDBAsAdmin.find(delq)}
-    catch(err) { console.log("ERROR: Could not find items/lists to delete:",err); return false;}
-    console.log("Found items/lists to delete:",foundDelDocs.docs.length);
-    for (let i = 0; i < foundDelDocs.docs.length; i++) {
-        let dbResp=null;
-        try { dbResp=await todosDBAsAdmin.destroy(foundDelDocs.docs[i]._id,foundDelDocs.docs[i]._rev)}
-        catch(err) {console.log("ERROR deleting list/item:",err);}        
-    }
-    console.log("STATUS: Finished deleting lists and items.");
-    console.log("STATUS: Creating default listgroups for all users");
-    const userq = { selector: { type: "user", name: {$exists: true}}, limit: await totalDocCount(usersDBAsAdmin)};
-    let foundUserDocs;
-    try {foundUserDocs = await usersDBAsAdmin.find(userq);}
-    catch(err) {console.log("ERROR: Could not find user list during schema update:",err); return false;}
-    console.log("STATUS: Found users to create listgroups: ", foundUserDocs.docs.length);
-    for (let i = 0; i < foundUserDocs.docs.length; i++) {
-        const foundUserDoc = foundUserDocs.docs[i];
-        const listgroupq = { selector: { type: "listgroup", listGroupOwner: foundUserDoc.name, default: true},
-                             limit: await totalDocCount(todosDBAsAdmin)};
-        let foundListGroupDocs = await todosDBAsAdmin.find(listgroupq);
-        if (foundListGroupDocs.docs.length == 0) {
-            console.log("STATUS: No default listgroup found for :",foundUserDoc.name," ... creating...");
-            let newCurDateStr = (new Date()).toISOString()
-            const newListGroupDoc = {
-                type: "listgroup", name: (foundUserDoc.name + " (default)"),
-                default: true, listGroupOwner: foundUserDoc.name, sharedWith: [], updatedAt: newCurDateStr
-            }
-            let dbResp = null;
-            try { dbResp = await todosDBAsAdmin.insert(newListGroupDoc)}
-            catch(err) { console.log("ERROR: Couldn't create new list group:",newListGroupDoc.name)
-                         updateSuccess = false;}
-        } else {
-            console.log("STATUS: Default List Group already exists for : ", foundUserDoc.name);
-        }
-    }
-    return updateSuccess;
-}    
-
-async function setSchemaVersion(updSchemaVersion) {
-    console.log("STATUS: Finished schema updates, updating database to :",updSchemaVersion);
-    let foundIDDoc = await getLatestDBUUIDDoc();
-    if (foundIDDoc == undefined) {
-        console.log("ERROR: Couldn't update database schema version record.");
-    } else {
-        foundIDDoc.schemaVersion = updSchemaVersion;
-        let dbResp = updateDBUUIDDoc(foundIDDoc);
-        if (dbResp == null) {console.log("ERROR: Couldn't update schema target version.")}
-        else {console.log("STATUS: Updated schema target version successfully.")}
-    }
-}
-
-async function checkAndUpdateSchema() {
-    console.log("STATUS: Current Schema Version:",schemaVersion," Target Version:",targetSchemaVersion);
-    if (schemaVersion === targetSchemaVersion) {
-        console.log("STATUS: At current schema version, skipping schema update");
-        return true;
-    }
-    if (schemaVersion < 2) {
-        console.log("STATUS: Updating schema to rev. 2: Changes for 'stocked at' indicator on item/list.");
-        let schemaUpgradeSuccess = await addStockedAtIndicatorToSchema();
-        if (schemaUpgradeSuccess) { schemaVersion = 2; await setSchemaVersion(schemaVersion);}
-    }
-    if (schemaVersion < 3) {
-        console.log("STATUS: Updating schema to rev. 3: Changes for restructuring/listgroups ");
-        let schemaUpgradeSuccess = await restructureListGroupSchema();
-        if (schemaUpgradeSuccess) { schemaVersion = 3; await setSchemaVersion(schemaVersion);}
-    }
-}
-
-async function createConflictsView() {
-    let viewFound=true; let existingView;
-    try {existingView = await todosDBAsAdmin.get("_design/"+conflictsViewID)}
-    catch(err) {viewFound = false;}
-    if (!viewFound) {
-        let viewCreated=true;
-        try {
-            await todosDBAsAdmin.insert({
-                "views": { "conflicts_view" : {
-                    "map": "function(doc) { if (doc._conflicts) { emit (doc._conflicts, null)}}"
-            }}},"_design/"+conflictsViewID)
-        }
-        catch(err) {console.log("ERROR: View not created:",{err}); viewCreated=false;}
-        console.log("STATUS: View created/ updated");
-    }
-}
-function isInteger(str) {
-    return /^\+?(0|[1-9]\d*)$/.test(str);
-}
-
-async function dbStartup() {
-    console.log("STATUS: Starting up auth server for couchdb...");
-    if (couchdbUrl == "") {console.log("ERROR: No environment variable for CouchDB URL"); return false;}
-    console.log("STATUS: Database URL: ",couchdbUrl);
-    if (couchDatabase == "") { console.log("ERROR: No CouchDatabase environment variable."); return false;}
-    console.log("STATUS: Using database: ",couchDatabase);
-    console.log("STATUS: Refresh token expires in ",refreshTokenExpires);
-    console.log("STATUS: Access token expires in ",accessTokenExpires);
-    await createDBIfNotExists();
-    await setDBSecurity();
-    try {todosDBAsAdmin = todosNanoAsAdmin.use(couchDatabase);}
-    catch(err) {console.log("ERROR: Could not open todo database:",err); return false;}
-    try {usersDBAsAdmin = usersNanoAsAdmin.use("_users");}
-    catch(err) {console.log("ERROR: Could not open users database:", err); return false;}
-    await addDBIdentifier();
-    await checkAndUpdateSchema();
-    await checkAndCreateContent();
-    await createConflictsView();
-    if (enableScheduling) {
-        if(isInteger(resolveConflictsFrequencyMinutes)) {
-            setInterval(() => {resolveConflicts()},60000*Number(resolveConflictsFrequencyMinutes));
-            console.log("STATUS: Conflict resolution scheduled every ",resolveConflictsFrequencyMinutes, " minutes.")
-            resolveConflicts();
-        } else {
-            console.log("ERROR: Invalid environment variable for scheduling conflict resolution -- not started.");
-        }
-        if (isInteger(expireJWTFrequencyMinutes)) {
-            setInterval(() => {expireJWTs()},60000*Number(expireJWTFrequencyMinutes));
-            console.log("STATUS: JWT expiry scheduled every ",expireJWTFrequencyMinutes," minutes.");
-            expireJWTs();
-        } else {
-            console.log("ERROR: Invalid environment variable for scheduling JWT expiry -- not started")
-        }
-    }
-}
-
-async function getUserDoc(username) {
-    const userResponse = {
-        error: false,
-        fullname: "",
-        email: "",
-        fullDoc: {}
-    }
-    let res = null;
-    try { res = await usersDBAsAdmin.get(couchUserPrefix+":"+username)}
-    catch(err) { console.log("ERROR GETTING USER:",err); userResponse.error= true }
-    if (!userResponse.error) {
-        userResponse.email = res.email;
-        userResponse.fullname = res.fullname;
-        userResponse.fullDoc = res;
-    }
-    return (userResponse);
-}
-
-async function generateJWT({ username, deviceUUID, timeString, includeRoles}) {
-    const alg = "HS256";
-    const secret = new TextEncoder().encode(couchKey);
-    const payload = {'sub': username, 'deviceUUID': deviceUUID};
-    if (includeRoles) { 
-        payload["_couchdb.roles"] =  [couchStandardRole];
-        payload["tokenType"] = "access"
-    } else {
-        payload["tokenType"] = "refresh"
-    }
-    const jwt = await new jose.SignJWT(payload)
-        .setProtectedHeader({ alg })
-        .setIssuedAt()
-        .setExpirationTime(timeString)
-        .sign(secret);  
-    return (jwt);
-}
-
-async function checkUserExists(req, res) {
-    const { username } = req.body;
-    let response = {
-        username: username,
-        userExists: false
-    }
-    let userResponse = await getUserDoc(username);
-    response.userExists = !userResponse.error;
-    return (response);
-}
-
-async function getUserByEmailDoc(email) {
-    const userResponse = {
-        error: false,
-        username: null,
-        fullname: null,
-        email: email,
-    }
-    const query={selector: {"email": {"$eq": email}}, limit: totalDocCount(usersDBAsAdmin)};
-    let res = null;
-    try { res = await usersDBAsAdmin.find(query);}
-    catch(err) { console.log("ERROR getting user by email:",err); userResponse.error= true }
-    if (!userResponse.error) {
-        if (res.docs.length > 0) {
-            userResponse.username = res.docs[0].name;
-            userResponse.email = res.docs[0].email;
-            userResponse.fullname = res.docs[0].fullname;
-        } else {
-            userResponse.error = true;
-        }
-    }
-    return (userResponse);
-}
-
-async function checkUserByEmailExists(req, res) {
-    const { email} = req.body;
-    let response = {
-        username: null,
-        fullname: null,
-        email: null,
-        userExists: true
-    }
-    let userResponse = await getUserByEmailDoc(email);
-    if (userResponse.error) { response.userExists = false;}
-    else {
-        response.username = userResponse.username;
-        response.email = userResponse.email;
-        response.fullname = userResponse.fullname;
-    }
-    return (response);
-}
-
-
-async function issueToken(req, res) {
-    const { username, password, deviceUUID } = req.body;
-    console.log("STATUS: issuing token for device ID:", JSON.stringify(deviceUUID));
-    let response = {
-        loginSuccessful: false,
-        email: "",
-        fullname: "",
-        loginRoles: [],
-        refreshJWT: "",
-        accessJWT: "",
-        couchdbUrl: process.env.COUCHDB_URL,
-        couchdbDatabase: process.env.COUCHDB_DATABASE
-    }
-    let loginResponse = await couchLogin(username,password);
-    if (!loginResponse.loginSuccessful) {
-         console.log("STATUS: Authentication failed for device: ",deviceUUID, " user: ",username);
-         return (response);
-        }     
-    let userDoc = await getUserDoc(username);
-    if (loginResponse.loginSuccessful && !(userDoc.error)) {
-         response.loginSuccessful = loginResponse.loginSuccessful;
-         response.loginRoles = loginResponse.loginRoles;
-         response.email = userDoc.email;
-         response.fullname = userDoc.fullname;
-         response.refreshJWT = await generateJWT({username: username, deviceUUID: deviceUUID, includeRoles:false, timeString: refreshTokenExpires});
-         response.accessJWT = await generateJWT({username: username, deviceUUID: deviceUUID, includeRoles:true, timeString: accessTokenExpires});
-     }
-     if (!userDoc.fullDoc.hasOwnProperty('refreshJWTs')) {
-        (userDoc.fullDoc as any).refreshJWTs = {};
-     }
-     (userDoc.fullDoc as any).refreshJWTs[deviceUUID] = response.refreshJWT;
-     try {let userUpd = usersDBAsAdmin.insert(userDoc.fullDoc);}
-     catch(err) {console.log("ERROR: Could not update user: ",username,":",err); response.loginSuccessful=false;}
-    return(response);
-
-}
-
-async function isValidToken(refreshJWT) {
-    let jwtResponse = null;
-    let returnValue = {
-        isValid: false,
-        protectedHeader: null,
-        payload: null,
-        error : null
-    };
-    try { jwtResponse = await jose.jwtVerify(refreshJWT, JWTKey); }
-    catch(err) { returnValue.isValid = false; returnValue.error = err;
-            return returnValue;}
-    if (jwtResponse.hasOwnProperty("payload") && jwtResponse.hasOwnProperty("protectedHeader")) {
-        if (jwtResponse.payload.hasOwnProperty("sub")) {
-            returnValue.isValid = true;
-            returnValue.protectedHeader = jwtResponse.protectedHeader;
-            returnValue.payload = jwtResponse.payload;
-        }
-    }        
-    return returnValue;
-}
-
-async function JWTMatchesUserDB(refreshJWT, deviceUUID, username) {
-    let userDoc: any = await getUserDoc(username);
-    if (userDoc.fullDoc.hasOwnProperty('refreshJWTs')) {
-        console.log("Refresh JWT matches the database:",userDoc.fullDoc.refreshJWTs[deviceUUID] == refreshJWT);    
-    }
-    if (userDoc.error) { return false;}
-    if (userDoc.fullDoc.name !== username) { return false;}
-    if (!userDoc.fullDoc.hasOwnProperty("refreshJWTs")) { return false;}
-    if (userDoc.fullDoc.refreshJWTs[deviceUUID] !== refreshJWT) { return false;}
-    return true;
-}
-
-async function invalidateToken(username, deviceUUID, invalidateAll) {
-    console.log("invalidating token now...");
-    let userDoc: any = await getUserDoc(username);
-    if (userDoc.error) { return false;}
-    if (userDoc.fullDoc.name !== username) { return false;}
-    if (!userDoc.fullDoc.hasOwnProperty("refreshJWTs")) { return false;}
-    if (invalidateAll) {
-        userDoc.fullDoc.refreshJWTs = {};
-    } else {
-        userDoc.fullDoc.refreshJWTs[deviceUUID] = {};
-    }    
-    try { let res = await usersDBAsAdmin.insert(userDoc.fullDoc); }
-    catch(err) { console.log("ERROR: problem invalidating token: ",err); return false; }
-    console.log("token now invalidated");
-    return true;
-}
-
-async function refreshToken(req, res) {
-    const { refreshJWT, deviceUUID } = req.body;
-    console.log("Refreshing token for deviceUUID:",deviceUUID);
-    // validate incoming refresh token
-    //      valid by signature and expiration date
-    //      matches most recent in user DB
-    // if valid then:
-    //       generate new access and refresh JWT
-    //       update userDB with current JWT
-    //       return access and refresh JWT
-    let status = 200;
-    let response = {
-        valid : false,
-        refreshJWT: "",
-        accessJWT: ""
-    }
-    const tokenDecode = await isValidToken(refreshJWT);
-    console.log("username of decoded token:",tokenDecode.payload.sub);
-    if (!tokenDecode.isValid) {
-        status = 403;
-        return({status, response});
-    }
-    if (tokenDecode.payload.deviceUUID !== deviceUUID) {
-        console.log("SECURITY: Attempt to use refresh token with mis-matched device UUIDs. Invalidating all JWTs for ",tokenDecode.payload.sub);
-        invalidateToken(tokenDecode.payload.sub,deviceUUID,true)
-        status = 403;
-        return({status, response});
-    }
-    if (! (await JWTMatchesUserDB(refreshJWT,deviceUUID, tokenDecode.payload.sub))) {
-        console.log("SECURITY: Login for user ",tokenDecode.payload.sub,"  didn't match stored database JWT. Invalidating this device.");
-        status = 403;
-        await invalidateToken(tokenDecode.payload.sub, deviceUUID, false);
-        return ({status, response});
-    }
-    response.valid = true;
-    response.refreshJWT = await generateJWT({username: tokenDecode.payload.sub, deviceUUID: deviceUUID, includeRoles: false, timeString: refreshTokenExpires});
-    response.accessJWT = await generateJWT({username: tokenDecode.payload.sub, deviceUUID: deviceUUID, includeRoles: true, timeString: accessTokenExpires});
-    let userResponse: any = await getUserDoc(tokenDecode.payload.sub);
-    userResponse.fullDoc.refreshJWTs[deviceUUID] = response.refreshJWT;
-    try {let update = await usersDBAsAdmin.insert(userResponse.fullDoc);}
-    catch(err) {console.log("ERROR: Could not update user(refresh token):",err); response.valid=false;}
-//    console.log("about to return from refresh Token:",{status},{response});
-    return ({status, response});
-}
-
-async function createNewUser(userObj, deviceUUID) {
-    console.log("In Create New User, deviceUUID is :",deviceUUID);
-    const createResponse = {
-        error: false,
-        idCreated: "",
-        refreshJWT: null,
-        accessJWT: null
-    }
-    let refreshJWTs = {};
-    if (deviceUUID !== "") {
-        let newJWT = await generateJWT({username: userObj.username,deviceUUID: deviceUUID,includeRoles: false, timeString: refreshTokenExpires});
-        refreshJWTs = { deviceUUID: newJWT}
-        createResponse.refreshJWT = newJWT;
-    }
-    const newDoc = {
-        name: userObj.username,
-        password: userObj.password,
-        email: userObj.email,
-        fullname: userObj.fullname,
-        roles: [couchStandardRole],
-        refreshJWTs: refreshJWTs,
-        type: "user"
-    }
-    if (deviceUUID !== "") {
-        createResponse.accessJWT = await generateJWT({username: userObj.username,deviceUUID: deviceUUID, includeRoles: true, timeString: accessTokenExpires});
-    }    
-    let res = null;
-    try { res = await usersDBAsAdmin.insert(newDoc,couchUserPrefix+":"+userObj.username); }
-    catch(err) { console.log("ERROR: problem creating user: ",err); createResponse.error= true }
-    if (!createResponse.error) {
-        createResponse.idCreated = res.id;
-    }
-    return (createResponse);
-}
-
-async function logout(req, res) {
-    const { refreshJWT, deviceUUID, username } = req.body;
-    console.log("logging out user: ", username, " for device: ",deviceUUID);
-    let userResponse: any = await getUserDoc(username);
-    userResponse.fullDoc.refreshJWTs[deviceUUID] = ""; 
-    let update = null;
-    try { update = await usersDBAsAdmin.insert(userResponse.fullDoc); res.sendStatus(200);}
-    catch(err) { console.log("ERROR: problem logging out user: ",err); res.sendStatus(404); }
-}
-
-async function authenticateJWT(req,res,next) {
-    const authHeader = req.headers.authorization;
-    if (authHeader) {
-        const token = authHeader.split(' ')[1];
-        const tokenDecode = await isValidToken(token);
-        if (!tokenDecode.isValid) {
-            return res.sendStatus(403);
-        }
-        req.username = tokenDecode.payload.sub;
-        next();
-    } else {
-        res.sendStatus(401);
-    }
-}
-
-async function updateUnregisteredFriends(req,email) {
-    const emailq = {
-        selector: { type: { "$eq": "friend" }, inviteEmail: { "$eq": email}},
-        limit: await totalDocCount(todosDBAsAdmin)
-    }
-    let foundFriendDocs;
-    try {foundFriendDocs =  await todosDBAsAdmin.find(emailq);}
-    catch(err) {console.log("ERROR: Could not find friend documents:",err); return false;}
-    let foundFriendDoc = undefined;
-//    if (foundFriendDocs.docs.length > 0) {foundFriendDoc = foundFriendDocs.docs[0]}
-    foundFriendDocs.docs.forEach(async (doc) => {
-        if (doc.friendStatus == "WAITREGISTER") {
-            doc.friendID2 = req.body.username;
-            doc.friendStatus = "PENDFROM1";
-            doc.updatedAt = (new Date()).toISOString();
-            let update2success=true;
-            try { await todosDBAsAdmin.insert(doc);} 
-            catch(e) {update2success = false;}
-        }
-    });
-}
-
-function isNothing(obj) {
-    if (obj == "" || obj == null || obj == undefined) {return (true)}
-    else {return (false)};
-}
-
-async function registerNewUser(req, res) {
-    const {username, password, email, fullname, deviceUUID} = req.body;
-    const registerResponse = {
-        invalidData: false,
-        userAlreadyExists: false,
-        createdSuccessfully: false,
-        idCreated: "",
-        refreshJWT: "",
-        accessJWT: "",
-        couchdbUrl: process.env.COUCHDB_URL,
-        couchdbDatabase: process.env.COUCHDB_DATABASE,
-        email: email,
-        fullname: fullname
-    }
-
-    if (isNothing(username) || isNothing(password) || isNothing(email) || isNothing(fullname)) {
-        registerResponse.invalidData = true;
-        return (registerResponse);
-    }
-
-    let userDoc = await getUserDoc(username);
-    if (!userDoc.error) {
-        registerResponse.userAlreadyExists = true;
-    } 
-    if (!registerResponse.userAlreadyExists)  {
-        let createResponse = await createNewUser({username: username, password: password, email: email, fullname: fullname}, deviceUUID);
-        registerResponse.createdSuccessfully = !createResponse.error;
-        registerResponse.idCreated = createResponse.idCreated;
-        registerResponse.refreshJWT = createResponse.refreshJWT;
-        registerResponse.accessJWT = createResponse.accessJWT;
-        let updateFriendResponse = await updateUnregisteredFriends(req,email)
-    }
-    return (registerResponse);
-}
-
-async function getUsersInfo(req, res) {
-    // input - json list of userIDs : userIDs: ["username1","username2"] -- should be _users ids 
-    //        without the org.couchdb.user prefix
-    // return - json array of objects:
-    //        [ {userID: "username1", email: "username1@gmail.com", fullName: "User 1"},
-    //          {userID: "username2", email: "username2@yahoo.com", fullName: "User 2"}]
-
-    const getResponse = {
-        error: false,
-        users: [],
-    }
-    if (isNothing(req.body?.userIDs)) {getResponse.error=true; return (getResponse)}
-    const requestData = { keys: [], include_docs: true }
-    req.body.userIDs.forEach(uid => { requestData.keys.push(couchUserPrefix+":"+uid) });
-    let userRes = null;
-    try { userRes = await usersDBAsAdmin.list(requestData);}
-    catch(err) { console.log("ERROR: problem retrieving users: ",err); getResponse.error= true }
-    if (!getResponse.error) {
-        userRes.rows.forEach(el => {
-            getResponse.users.push({name: el?.doc?.name, email: el?.doc?.email, fullname: el?.doc?.fullname})
-        });
-    }
-    return(getResponse);
-}
-
-async function createAccountUIGet(req, res) {
-    // input - query parameter - uuid
-    // creates a form pre-filled with email address (cannot change), username,full name, and password
-    // on submit (to different endpoint?),
-    // check if username or email already exists, if so, error out,
-    // otherwise reqister new user
-
-    let respObj = {
-        uuid: req.query.uuid,
-        email: "testemail",
-        fullname: "",
-        username: "",
-        password: "",
-        passwordverify: "",
-        formError: "",
-        disableSubmit: false,
-        createdSuccessfully: false
-    }
-    const uuidq = {
-        selector: { type: { "$eq": "friend" }, inviteUUID: { "$eq": req.query.uuid}},
-        fields: [ "friendID1","friendID2","inviteUUID","inviteEmail","friendStatus"],
-        limit: await totalDocCount(todosDBAsAdmin)
-    }
-    let foundFriendDocs;
-    try {foundFriendDocs =  await todosDBAsAdmin.find(uuidq);}
-    catch(err) {console.log("ERROR: Could not find friend documents:",err);
-                respObj.formError="Database Error Encountered";
-                return respObj;}
-    let foundFriendDoc;
-    if (foundFriendDocs.docs.length > 0) {foundFriendDoc = foundFriendDocs.docs[0]}
-    else {
-        respObj.formError = "Registration ID not found. Cannot complete registration process."
-        respObj.disableSubmit = true;
-        return (respObj);
-    };
-    respObj.email = foundFriendDoc.inviteEmail;
-
-    if (foundFriendDoc.friendStatus != "WAITREGISTER") {
-        respObj.formError = "Registration already completed. Please login to the app directly."
-        respObj.disableSubmit = true;
-        return (respObj);
-    }
-
-    return(respObj);
-
-}
-
-async function getFriendDocByUUID(uuid) {
-    const uuidq = {
-        selector: { type: { "$eq": "friend" }, inviteUUID: { "$eq": uuid}},
-        limit: await totalDocCount(todosDBAsAdmin)
-    }
-    let foundFriendDocs;
-    try {foundFriendDocs =  await todosDBAsAdmin.find(uuidq);}
-    catch(err) {console.log("ERROR: Could not find friend documents:", err); return null;};
-    let foundFriendDoc;
-    if (foundFriendDocs.docs.length > 0) {foundFriendDoc = foundFriendDocs.docs[0]}
-    return(foundFriendDoc);
-}
-
-async function createAccountUIPost(req,res) {
-
-    let respObj = {
-        uuid: req.body.uuid,
-        email: req.body.email,
-        username: (req.body.username == undefined) ? "" : req.body.username,
-        fullname: (req.body.fullname == undefined) ? "" : req.body.fullname,
-        password: (req.body.password == undefined) ? "" : req.body.password,
-        passwordverify: (req.body.passwordverify == undefined) ? "" : req.body.passwordverify,
-        refreshjwt: "",
-        formError: "",
-        disableSubmit: false,
-        createdSuccessfully: false
-    }
-
-    if (req.body.fullname.length < 2 ) {
-        respObj.formError = "Please enter a full name 3 characters or longer";
-        return (respObj);} 
-    if (!fullnamePatternValidation(req.body.fullname)) {
-        respObj.formError = "Please enter a valid full name";
-        return (respObj);}
-    if (req.body.username.length < 5 ) {
-        respObj.formError = "Please enter a username 5 characters or longer";
-        return (respObj); } 
-    if (!usernamePatternValidation(req.body.username)) {
-        respObj.formError = "Please enter a valid username";
-        return (respObj); }
-    if (req.body.password.length < 5 ) {
-        respObj.formError = "Please enter a password 5 characters or longer";
-        return (respObj);} 
-    if (req.body.password != req.body.passwordverify) {
-        respObj.formError = "Passwords do not match";
-        return (respObj);}
-    let foundUserDoc; let userAlreadyExists=true;
-    try {
-        foundUserDoc =  await usersDBAsAdmin.get(couchUserPrefix+":"+req.body.username);}
-    catch(e) { userAlreadyExists=false;}    
-    if (userAlreadyExists) {
-        respObj.formError = "Username already exists, plase choose a new one";
-        return(respObj);
-    }
-
-    // create user doc
-    const userObj = {
-        username: req.body.username,
-        fullname: req.body.fullname,
-        email: req.body.email,
-        password: req.body.password
-    }
-
-    let userIDRes = await createNewUser(userObj,"");
-    respObj.refreshjwt = userIDRes.refreshJWT;
-
-    // change friend doc to registered
-    let foundFriendDoc = await getFriendDocByUUID(req.body.uuid);
-    if (foundFriendDoc!=undefined) {
-        foundFriendDoc.friendID2 = req.body.username;
-        foundFriendDoc.friendStatus = "PENDFROM1";
-        foundFriendDoc.updatedAt = (new Date()).toISOString();
-        let updateSuccessful = true;
-        try { await todosDBAsAdmin.insert(foundFriendDoc); }
-        catch(e) {updateSuccessful = false;}
-    }
-
-    const emailq = {
-        selector: { type: { "$eq": "friend" }, inviteEmail: { "$eq": req.body.email},
-                    limit: totalDocCount(todosDBAsAdmin)}
-    }
-    let foundFriendDocs;
-    try {foundFriendDocs =  await todosDBAsAdmin.find(emailq);}
-    catch(err) {console.log("ERROR: Could not find friend by email:",err); 
-                respObj.formError="Database error finding friend by email";
-                return respObj;}
-    foundFriendDoc = undefined;
-    if (foundFriendDocs.docs.length > 0) {foundFriendDoc = foundFriendDocs.docs[0]}
-    foundFriendDocs.docs.forEach(async (doc) => {
-        if ((doc.inviteUUID != req.body.uuid) && (doc.friendStatus == "WAITREGISTER")) {
-            doc.friendID2 = req.body.username;
-            doc.friendStatus = "PENDFROM1";
-            doc.updatedAt = (new Date()).toISOString();
-            let update2Success=true;
-            try { await todosDBAsAdmin.insert(doc);} 
-            catch(e) {update2Success = false;}
-        }
-    });
-
-    respObj.createdSuccessfully = true;
-    return(respObj);
-}
-
-async function triggerRegEmail(req, res) {
-    const triggerResponse = {
-        emailSent : false
-    }
-    const {uuid} = req.body;
-    if (isNothing(uuid)) {return (triggerResponse);}
-    let foundFriendDoc = await getFriendDocByUUID(req.body.uuid);
-    if (foundFriendDoc == undefined) {return triggerResponse;};
-    let userDoc = await getUserDoc(foundFriendDoc.friendID1);
-    if (userDoc.error) {return triggerResponse};
-    
-    let transport = nodemailer.createTransport(smtpOptions);
-    transport.verify(function (error,success) {
-        if (error) {return triggerResponse}
-    })
-
-    let confURL = groceryAPIUrl + "/createaccountui?uuid="+foundFriendDoc.inviteUUID;
-
-    const message = {
-        from: smtpFrom,
-        to: foundFriendDoc.inviteEmail,
-        subject: "User Creation request from Groceries",
-        text: userDoc.fullname+" has requested to share lists with you on the Groceries App. Please use the link to register for an account: "+confURL+ " . Once registered, visit "+groceryUrl+" to use the app."
-    }
-
-    transport.sendMail(message, function (error, success) {
-        if (!error) {return triggerResponse}
-    });
-    triggerResponse.emailSent = true;
-
-    return (triggerResponse);
-}
-
-async function resetPassword(req, res) {
-    const resetResponse = {
-        emailSent : false
-    }
-    const {username} = req.body;
-
-    let userDoc = await getUserDoc(username);
-    if (userDoc.error) {return resetResponse};
-
-    let transport = nodemailer.createTransport(smtpOptions);
-    transport.verify(function (error,success) {
-        if (error) {return resetResponse}
-    })
-
-    let resetURL = groceryAPIUrl + "/resetpasswordui?username="+encodeURIComponent(username);
-    const message = {
-        from: smtpFrom,
-        to: userDoc.email,
-        subject: "Password Reset Request from Groceries",
-        text: userDoc.fullname+" has requested to reset their password. If you did not request this reset, please validate the security of this account. Please use the link to reset your password: "+resetURL+ " . Once reset, visit "+groceryUrl+" to use the app or login on your mobile device.."
-    }
-
-    transport.sendMail(message, function (error, success) {
-        if (!error) {return resetResponse}
-    });
-    resetResponse.emailSent = true;
-    return resetResponse;
-}    
-
-async function resetPasswordUIGet(req, res) {
-    let respObj = {
-        email: "",
-        username: req.query.username,
-        password: "",
-        passwordverify: "",
-        formError: "",
-        disableSubmit: false,
-        resetSuccessfully: false
-    }
-    
-    let userDoc = await getUserDoc(req.query.username);
-    if (userDoc.error) {
-        respObj.formError = "Cannot locate user name "+req.query.username;
-        respObj.disableSubmit = true;
-        return respObj
-    }
-    respObj.email = userDoc.email;
-    return respObj;
-}
-
-async function resetPasswordUIPost(req, res) {
-    let respObj = {
-        username: req.body.username,
-        password: (req.body.password == undefined) ? "" : req.body.password,
-        passwordverify: (req.body.passwordverify == undefined) ? "" : req.body.passwordverify,
-        email: (req.body.email),
-        formError: "",
-        disableSubmit: false,
-        resetSuccessfully: false
-    }
-    if (req.body.password.length < 5 ) {
-        respObj.formError = "Please enter a password 5 characters or longer";
-        return (respObj);
-    } 
-    if (req.body.password != req.body.passwordverify) {
-        respObj.formError = "Passwords do not match";
-        return (respObj);
-    }
-
-    let userDoc = await getUserDoc(req.body.username);
-    if (!userDoc.error) {
-        let newDoc=cloneDeep(userDoc.fullDoc);
-        newDoc.password=req.body.password;
-        let newDocFiltered = _.pick(newDoc,['_id','_rev','name','email','fullname','roles','type','password','refreshJWTs']);
-        try {let docupdate = await usersDBAsAdmin.insert(newDocFiltered);}
-        catch(err) {console.log("ERROR: Couldn't update user/reset password:",err);
-                    respObj.formError="Database error resetting password";
-                    respObj.resetSuccessfully=false;
-                    return respObj;}
-    }
-    respObj.resetSuccessfully = true;
-    return(respObj);
-}
-
-async function resolveConflicts() {
-    let conflicts;
-    try {conflicts = await todosDBAsAdmin.view(conflictsViewID,conflictsViewName);}
-    catch(err) {console.log("ERROR: Couldn't create conflicts view:",err); return false;}
-    console.log("STATUS: Resolving all conflicts started...");
-    let resolveFailure=false;
-    if (conflicts.rows?.length <= 0) {console.log("STATUS: no conflicts found"); return};
-    outerloop: for (let i = 0; i < conflicts.rows.length; i++) {
-        const conflict = conflicts.rows[i];
-        let curWinner;
-        try { curWinner = await todosDBAsAdmin.get(conflict.id, {conflicts: true});}
-        catch(err) { console.log("ERROR: Error resolving conflicts:",err); resolveFailure = true;}
-        if (resolveFailure) {continue};
-        let latestDocTime = curWinner.updatedAt; 
-        let latestIsCurrentWinner = true;
-        let latestDoc = curWinner
-        let bulkObj = { docs: []};
-        let logObj = { type: "conflictlog", docType: curWinner.type, winner: {}, losers: [], updatedAt: ""};
-        for (let j = 0; j < curWinner._conflicts.length; j++) {
-            const losingRev = curWinner._conflicts[j];
-            let curLoser;
-            try { curLoser = await todosDBAsAdmin.get(conflict.id,{ rev: losingRev})}
-            catch(err) {console.log("ERROR: Error resolving conflicts:",err); resolveFailure=true;}
-            if (resolveFailure) {continue outerloop};
-            if (curLoser.updatedAt >= latestDocTime) {
-                latestIsCurrentWinner = false;
-                latestDocTime = curLoser.updatedAt;
-                latestDoc = curLoser;
-            } else {
-                bulkObj.docs.push({_id: curLoser._id, _rev: losingRev ,_deleted: true})
-                logObj.losers.push(curLoser);
-            }
-        }
-        let resolvedTime = (new Date()).toISOString();
-        latestDoc.updatedAt = resolvedTime;
-        logObj.updatedAt = resolvedTime;
-        bulkObj.docs.push(latestDoc);
-        logObj.winner = latestDoc;
-        if (!latestIsCurrentWinner) {
-            bulkObj.docs.push({_id: curWinner._id, _rev: curWinner._rev, _deleted: true});
-            logObj.losers.push(curWinner);
-        }
-        let bulkResult;
-        try { bulkResult = await todosDBAsAdmin.bulk(bulkObj) }
-        catch(err) {console.log("ERROR: Error updating bulk docs on conflict resolve"); resolveFailure=true;}
-        console.log("STATUS: Bulk Update to resolve doc id : ",conflict.id, " succeeded");
-        let logResult;
-        try { logResult = await todosDBAsAdmin.insert(logObj)}
-        catch(err) { console.log("ERROR: creating conflict log document failed: ",err)};
-    }
-}
-
-async function expireJWTs() {
-    console.log("STATUS: Checking for expired JWTs");
-    let userDocs;
-    try {userDocs = await usersDBAsAdmin.list({include_docs: true});}
-    catch(err) {console.log("ERROR: Could not find user records to expire JWTs:",err); return false;}
-    for (let i = 0; i < userDocs.rows.length; i++) {
-        const userDoc = userDocs.rows[i].doc;
-        if (userDoc.hasOwnProperty("refreshJWTs")) {
-            let updateJWTs = {};
-            for (const [device,jwt] of Object.entries(userDoc.refreshJWTs)) {
-                if (isEmpty(jwt)) {continue;}
-                let jwtVerify = await isValidToken(jwt);
-                if (jwtVerify.isValid) {updateJWTs[device] = jwt}
-            }
-            if (!isEqual(updateJWTs,userDoc.refreshJWTs)) {
-                userDoc.refreshJWTs=updateJWTs;
-                try { let response=usersDBAsAdmin.insert(userDoc); }
-                catch(err) {console.log("ERROR updating JWTs for user ", userDoc._id, err)}
-                console.log("STATUS: Expired JWTs for user ",userDoc._id);
-            } 
-        }
-    }
-    console.log("STATUS: Finished checking for expired JWTs");
-}
-
-async function triggerResolveConflicts(req,res) {
-    let respObj = {
-        triggered: true
-    };
-    resolveConflicts()
-    return respObj;
-}
-
-async function compactDB() {
-    todosNanoAsAdmin.db.compact(couchDatabase);
-}
-
-async function triggerDBCompact(req,res) {
-    let respObj = {
-        triggered: true
-    };
-    compactDB();
-    return respObj;
-}
-
-export {
-    issueToken,
-    refreshToken,
-    checkUserExists,
-    checkUserByEmailExists,
-    registerNewUser,
-    dbStartup,
-    getUsersInfo,
-    createAccountUIGet,
-    createAccountUIPost,
-    triggerRegEmail,
-    resetPassword,
-    resetPasswordUIGet,
-    resetPasswordUIPost,
-    triggerResolveConflicts,
-    triggerDBCompact,
-    authenticateJWT,
-    logout
-
-}
